% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{add_missing_schema_columns}
\alias{add_missing_schema_columns}
\title{Add missing schema columns with appropriate defaults}
\usage{
add_missing_schema_columns(df)
}
\arguments{
\item{df}{Dataframe to enhance}
}
\value{
Dataframe with all schema columns
}
\description{
Add missing schema columns with appropriate defaults
}
