% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_record_ids}
\alias{add_record_ids}
\title{Create record IDs for a batch of records (internal)}
\usage{
add_record_ids(interactions, author_lastname, publication_year)
}
\arguments{
\item{interactions}{Dataframe of records}

\item{author_lastname}{Author lastname for ID generation}

\item{publication_year}{Publication year for ID generation}
}
\value{
Dataframe with record_id column added
}
\description{
Create record IDs for a batch of records (internal)
}
\keyword{internal}
