% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoextract-package.R
\docType{package}
\name{ecoextract-package}
\alias{ecoextract}
\alias{ecoextract-package}
\title{ecoextract: Ecological Data Extraction and Refinement}
\description{
Structured ecological data extraction and refinement from scientific literature. Provides functions for OCR processing, extracting structured data from documents, and refining extracted data using LLMs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/n8layman/ecoextract}
  \item \url{https://n8layman.github.io/ecoextract/}
  \item Report bugs at \url{https://github.com/n8layman/ecoextract/issues}
}

}
\author{
\strong{Maintainer}: Nathan Layman \email{nathan.layman@gmail.com} (\href{https://orcid.org/0000-0003-2238-6584}{ORCID})

Authors:
\itemize{
  \item Nathan Layman \email{nathan.layman@gmail.com} (\href{https://orcid.org/0000-0003-2238-6584}{ORCID})
}

Other contributors:
\itemize{
  \item Valentine Herrmann [contributor]
  \item Cadhla Firt [contributor]
  \item Amanda Liew [contributor]
  \item Aja Sherman [contributor]
}

}
\keyword{internal}
