% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrich_publication_metadata}
\alias{enrich_publication_metadata}
\title{Publication Metadata Enrichment Functions}
\usage{
enrich_publication_metadata(doi = NULL, title = NULL, authors = NULL)
}
\arguments{
\item{doi}{DOI to lookup}

\item{title}{Publication title (used if DOI not available)}

\item{authors}{Author names (used if DOI not available)}
}
\value{
List with enriched metadata
}
\description{
Enrich publication metadata using CrossRef API
Enrich publication metadata using CrossRef
}
