% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{extract_authors_from_crossref}
\alias{extract_authors_from_crossref}
\title{Extract authors array from CrossRef author data}
\usage{
extract_authors_from_crossref(crossref_authors)
}
\arguments{
\item{crossref_authors}{Author data from CrossRef}
}
\value{
JSON string of author array
}
\description{
Extract authors array from CrossRef author data
}
