% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{extract_first_author_lastname}
\alias{extract_first_author_lastname}
\title{Extract first author lastname from CrossRef data}
\usage{
extract_first_author_lastname(crossref_authors)
}
\arguments{
\item{crossref_authors}{Author data from CrossRef}
}
\value{
Character string of first author lastname
}
\description{
Extract first author lastname from CrossRef data
}
