% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{generate_columns_from_ellmer_schema}
\alias{generate_columns_from_ellmer_schema}
\title{Generate SQL column definitions from ellmer schema}
\usage{
generate_columns_from_ellmer_schema(ellmer_schema)
}
\arguments{
\item{ellmer_schema}{ellmer schema object}
}
\value{
Character string with SQL column definitions
}
\description{
Generate SQL column definitions from ellmer schema
}
