% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{generate_columns_from_json_schema}
\alias{generate_columns_from_json_schema}
\title{Generate SQL column definitions from JSON schema}
\usage{
generate_columns_from_json_schema(schema_json_list)
}
\arguments{
\item{schema_json_list}{Parsed JSON schema as list}
}
\value{
Character string with SQL column definitions
}
\description{
Generate SQL column definitions from JSON schema
}
\keyword{internal}
