% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{generate_occurrence_id}
\alias{generate_occurrence_id}
\title{Generate occurrence ID for interaction}
\usage{
generate_occurrence_id(author_lastname, publication_year, sequence_number = 1)
}
\arguments{
\item{author_lastname}{Author surname}

\item{publication_year}{Publication year}

\item{sequence_number}{Sequence number for this interaction}
}
\value{
Character occurrence ID
}
\description{
Generate occurrence ID for interaction
}
