% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{generate_record_id}
\alias{generate_record_id}
\title{Generate record IDs (internal)}
\usage{
generate_record_id(prefix, sequence_number = 1)
}
\arguments{
\item{prefix}{Record ID prefix for the document, from \code{build_record_id_prefix()}}

\item{sequence_number}{Sequence number(s) of records within the document}
}
\value{
Character vector of record IDs
}
\description{
Generate record IDs (internal)
}
\keyword{internal}
