% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_existing_records}
\alias{get_existing_records}
\title{Get existing interactions from database}
\usage{
get_existing_records(document_id, db_conn)
}
\arguments{
\item{document_id}{Document ID to retrieve interactions for}

\item{db_conn}{Database connection}
}
\value{
Dataframe with existing interactions, or NA if none found
}
\description{
Get existing interactions from database
}
