% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata_context}
\alias{get_metadata_context}
\title{Get metadata context template (internal)}
\usage{
get_metadata_context(context_file = NULL)
}
\arguments{
\item{context_file}{Optional path to custom context template file}
}
\value{
Character string with context template
}
\description{
Get metadata context template (internal)
}
\keyword{internal}
