% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata_fields}
\alias{get_metadata_fields}
\title{Get field definitions from a metadata schema (internal)}
\usage{
get_metadata_fields(schema_list)
}
\arguments{
\item{schema_list}{Parsed metadata JSON schema}
}
\value{
Named list of field definitions
}
\description{
Get field definitions from a metadata schema (internal)
}
\keyword{internal}
