% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_record_columns_sql}
\alias{get_record_columns_sql}
\title{Get record table column definitions as SQL}
\usage{
get_record_columns_sql(schema_json_list = NULL)
}
\arguments{
\item{schema_json_list}{Optional parsed JSON schema to generate columns from}
}
\value{
Character string with column definitions
}
\description{
Get record table column definitions as SQL
}
