% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{get_required_columns}
\alias{get_required_columns}
\title{Get required columns from basic schema}
\usage{
get_required_columns()
}
\value{
Character vector of required column names
}
\description{
Get required columns from basic schema
}
