% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_schema_array_fields}
\alias{get_schema_array_fields}
\title{Get array field names from schema definition}
\usage{
get_schema_array_fields(schema_list)
}
\arguments{
\item{schema_list}{Parsed JSON schema (from jsonlite::fromJSON)}
}
\value{
Character vector of array field names
}
\description{
Identifies fields defined as arrays in the JSON schema.
}
\keyword{internal}
