% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{get_schema_columns}
\alias{get_schema_columns}
\title{Get column names from basic schema}
\usage{
get_schema_columns()
}
\value{
Character vector of column names
}
\description{
Get column names from basic schema
}
