% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{get_schema_types}
\alias{get_schema_types}
\title{Get database column types from basic schema (internal)}
\usage{
get_schema_types()
}
\value{
Named vector of SQL types
}
\description{
Get database column types from basic schema (internal)
}
\keyword{internal}
