% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{is_forced}
\alias{is_forced}
\title{Check if a document should be forced to reprocess}
\usage{
is_forced(force_param, document_id)
}
\arguments{
\item{force_param}{NULL, TRUE, or integer vector of document_ids}

\item{document_id}{The document ID to check}
}
\value{
logical
}
\description{
Check if a document should be forced to reprocess
}
\keyword{internal}
