% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{json_schema_to_ellmer_type_metadata}
\alias{json_schema_to_ellmer_type_metadata}
\title{Convert metadata JSON schema to native ellmer type specification}
\usage{
json_schema_to_ellmer_type_metadata(schema_path)
}
\arguments{
\item{schema_path}{Path to metadata JSON schema file}
}
\value{
Ellmer type specification object
}
\description{
Converts the metadata schema to ellmer's native type functions for
proper dataframe conversion
}
\keyword{internal}
