% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{json_type_to_sql}
\alias{json_type_to_sql}
\title{Map a JSON type to an SQLite column type (internal)}
\usage{
json_type_to_sql(json_type)
}
\arguments{
\item{json_type}{JSON type name}
}
\value{
SQL type string
}
\description{
Arrays and objects are stored as JSON text.
}
\keyword{internal}
