% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{list_prompts}
\alias{list_prompts}
\title{List all available prompts in package}
\usage{
list_prompts()
}
\value{
Character vector of available prompt files
}
\description{
List all available prompts in package
}
