% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_env}
\alias{load_env}
\title{Internal utility functions
Load environment variables from .env files}
\usage{
load_env()
}
\value{
NULL (invisibly). Called for side effects.
}
\description{
Loads environment variables from any file starting with `.env`. This allows
user-specific options to be set in `.env_user` (which is .gitignored), and
to have both encrypted and non-encrypted .env files.
}
