% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_message}
\alias{log_message}
\title{Simple logging function}
\usage{
log_message(message, level = "INFO")
}
\arguments{
\item{message}{Message to log}

\item{level}{Log level (INFO, WARNING, ERROR)}
}
\description{
Simple logging function
}
