% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinement.R
\name{merge_refinements}
\alias{merge_refinements}
\title{Merge refined data back into original interactions}
\usage{
merge_refinements(original_interactions, refined_interactions)
}
\arguments{
\item{original_interactions}{Dataframe of original interactions}

\item{refined_interactions}{Dataframe of refined interactions}
}
\value{
Dataframe with merged refinements
}
\description{
Merge refined data back into original interactions
}
