% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{migration_col_defs}
\alias{migration_col_defs}
\title{Build column definition strings for migration DDL from PRAGMA table_info output}
\usage{
migration_col_defs(info, overrides = list())
}
\arguments{
\item{info}{Data frame from PRAGMA table_info}

\item{overrides}{Named list mapping column name to full definition string}
}
\description{
Build column definition strings for migration DDL from PRAGMA table_info output
}
\keyword{internal}
