% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_occurrence_ids}
\alias{add_occurrence_ids}
\title{Create occurrence IDs for a batch of interactions}
\usage{
add_occurrence_ids(interactions, author_lastname, publication_year)
}
\arguments{
\item{interactions}{Dataframe of interactions}

\item{author_lastname}{Author lastname for ID generation}

\item{publication_year}{Publication year for ID generation}
}
\value{
Dataframe with occurrence_id column added
}
\description{
Create occurrence IDs for a batch of interactions
}
