% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{generate_record_id}
\alias{generate_record_id}
\title{Generate record ID for a record (internal)}
\usage{
generate_record_id(author_lastname, publication_year, sequence_number = 1)
}
\arguments{
\item{author_lastname}{Author surname}

\item{publication_year}{Publication year}

\item{sequence_number}{Sequence number for this record}
}
\value{
Character record ID
}
\description{
Generate record ID for a record (internal)
}
\keyword{internal}
