% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_db_stats}
\alias{get_db_stats}
\title{Get database statistics}
\usage{
get_db_stats(db_path)
}
\arguments{
\item{db_path}{Path to database file}
}
\value{
List with database statistics
}
\description{
Get database statistics
}
