% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinement.R
\name{merge_refinements}
\alias{merge_refinements}
\title{Merge refined data back into original records (internal)}
\usage{
merge_refinements(original_records, refined_records)
}
\arguments{
\item{original_records}{Dataframe of original records}

\item{refined_records}{Dataframe of refined records}
}
\value{
Dataframe with merged refinements
}
\description{
Merge refined data back into original records (internal)
}
\keyword{internal}
