% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocr.R
\name{perform_ocr}
\alias{perform_ocr}
\title{OCR Functions}
\usage{
perform_ocr(pdf_file)
}
\arguments{
\item{pdf_file}{Path to PDF}
}
\value{
List with markdown content and raw result
}
\description{
Functions for performing OCR on PDF documents
Perform OCR on PDF
}
\keyword{internal}
