% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{save_metadata_to_db}
\alias{save_metadata_to_db}
\title{Save publication metadata to existing document (internal)}
\usage{
save_metadata_to_db(document_id, db_conn, metadata = list())
}
\arguments{
\item{document_id}{Document ID}

\item{db_conn}{Database connection}

\item{metadata}{List with metadata fields (title, first_author_lastname, publication_year, doi, journal)}
}
\value{
Document ID
}
\description{
Save publication metadata to existing document (internal)
}
\keyword{internal}
