% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{save_records_to_db}
\alias{save_records_to_db}
\title{Save records to EcoExtract database}
\usage{
save_records_to_db(db_path, document_id, interactions_df, metadata = list())
}
\arguments{
\item{db_path}{Path to database file}

\item{document_id}{Document ID}

\item{interactions_df}{Dataframe of records}

\item{metadata}{Processing metadata}
}
\value{
TRUE if successful
}
\description{
Save records to EcoExtract database
}
